Change pixel datatype from double to float in window array

+++ include/graphics.h
@@ -43,7 +43,7 @@
 void init_gl(int argc, char** argv);
 
 /**
- * @brief Function: void create_window(int no_img, int width, int height, double * pixels, int * labels, win_type_e window_type)
+ * @brief Function: void create_window(int no_img, int width, int height, float * pixels, int * labels, win_type_e window_type)
  *
  * \param no_img: number of images in pixels array
  * \param width: width of the window.
@@ -54,7 +54,7 @@ void init_gl(int argc, char** argv);
  *
  * Add window to display elements
  */
-void create_window(int no_img, int width, int height, double * pixels, int * labels, win_type_e window_type);
+void create_window(int no_img, int width, int height, float * pixels, int * labels, win_type_e window_type);
 
 /**
  * @brief Function: void wrapper_dataset_cb()
